+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.0.2
+
+- only allow upright landings
+- added tiny delay to rerail for smoother feel
+- added more settings
+
 # 0.0.1
 
 Initial version.